+++ coordinate.h
@@ -15,7 +15,7 @@ public:
     Coordinate* previous;
     
     Coordinate();
-    Coordinate(int cost, int first, int second);
+    Coordinate(double cost, int first, int second);
     ~Coordinate();
     double getCost();
     void setCost(double cost);
